% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyCSV.R
\name{flyCSV}
\alias{flyCSV}
\title{Load a Matrix}
\usage{
flyCSV(df, filename = NA, browser = getOption("browser"), row.names = FALSE)
}
\arguments{
\item{df}{Object do be written. Usually data frame}

\item{filename}{a filename to save the CSV. If not provided, a temporary file will
be created}

\item{browser}{a non-empty character string giving the name of the program}

\item{rownames}{creates a row name column at the beginnig. Usually rows are named as the line number
to be used as the HTML browser.}
}
\value{
df Returns the original object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
