% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyCSV.R
\name{flyCSV}
\alias{flyCSV}
\title{Open your data as CSV}
\usage{
flyCSV(
  df,
  filename = NA,
  browser = getOption("browser"),
  row.names = FALSE,
  encoding = ""
)
}
\arguments{
\item{df}{Object do be written. Usually data frame.}

\item{filename}{a file name/path to save the CSV to. If provided, the file
will not be deleted afterwards.}

<<<<<<< HEAD
\item{browser}{a non-empty character string giving the name of the program
=======
\item{browser}{a non-empty character string giving the name of the program}

\item{encoding}{File encoding}

\item{rownames}{creates a row name column at the beginnig. Usually rows are named as the line number
>>>>>>> main
to be used as the HTML browser.}

\item{row.names}{Whether to include row names on the first column. TRUE or FALS.}

\item{encoding}{File encoding. Passes to write.csv}
}
\value{
df Returns the original object
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
